Fully implemented oceanforecast api call

+++ app/src/main/java/no/uio/ifi/in2000/testgit/model/oceanforecast/OceanDetails.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class OceanDetails (
 
-  @SerializedName("sea_surface_wave_from_direction" ) var seaSurfaceWaveFromDirection : Int?    = null,
+  @SerializedName("sea_surface_wave_from_direction" ) var seaSurfaceWaveFromDirection : Double?    = null,
   @SerializedName("sea_surface_wave_height"         ) var seaSurfaceWaveHeight        : Double? = null,
   @SerializedName("sea_water_speed"                 ) var seaWaterSpeed               : Double? = null,
   @SerializedName("sea_water_temperature"           ) var seaWaterTemperature         : Double? = null,
